+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/logging/LogManager.scala
@@ -44,3 +44,45 @@ abstract class LogManager {
   def terminate(): Unit
 
 }
+
+class EmptyLogManagerImpl(
+    networkCommunicationActor: NetworkCommunicationActor.NetworkSenderActorRef
+) extends LogManager {
+  override def setupWriter(logWriter: DeterminantLogStorage.DeterminantLogWriter): Unit = {}
+
+  override def getDeterminantLogger: DeterminantLogger = new EmptyDeterminantLogger()
+
+  override def sendCommitted(
+      sendRequest: NetworkCommunicationActor.SendRequest
+  ): Unit = {
+    networkCommunicationActor ! sendRequest
+  }
+
+  override def terminate(): Unit = {}
+}
+
+class LogManagerImpl(
+    networkCommunicationActor: NetworkCommunicationActor.NetworkSenderActorRef
+) extends LogManager {
+
+  private val determinantLogger = new DeterminantLoggerImpl()
+
+  private var writer: AsyncLogWriter = _
+
+  def setupWriter(logWriter: DeterminantLogWriter): Unit = {
+    writer = new AsyncLogWriter(networkCommunicationActor, logWriter)
+    writer.start()
+  }
+
+  def getDeterminantLogger: DeterminantLogger = determinantLogger
+
+  def sendCommitted(sendRequest: SendRequest): Unit = {
+    writer.putDeterminants(determinantLogger.drainCurrentLogRecords())
+    writer.putOutput(sendRequest)
+  }
+
+  def terminate(): Unit = {
+    writer.terminate()
+  }
+
+}
